+++ poly/src/domain/radix2.rs
@@ -193,10 +193,10 @@ impl<F: FftField> EvaluationDomain<F> for Radix2EvaluationDomain<F> {
         tau.pow(&[self.size]) - F::one()
     }
 
-    /// Returns the ith element of the domain, where elements are ordered by
+    /// Returns the `i`-th element of the domain, where elements are ordered by
     /// their power of the generator which they correspond to.
-    /// e.g. the ith element is g^i
-    fn get_element(&self, i: usize) -> F {
+    /// e.g. the `i`-th element is g^i
+    fn element(&self, i: usize) -> F {
         // TODO: Consider precomputed exponentiation tables if we need this to be faster.
         self.group_gen.pow(&[i as u64])
     }